Add API calls and error modal
Multiple changes have been made regarding the handling of "todos". The dummy data has been replaced with API calls to a backend server, allowing for the create, read, update and delete operations. The API responses are also being checked to ensure they are as expected. A modal for displaying errors was also added to inform the user whenever something goes wrong with these operations.