Add docker-entrypoint.sh and update port configuration in compose.yaml
Created a new shell script, docker-entrypoint.sh, to streamline the process of starting the server and migrating with Alembic. In addition, the port configuration in compose.yaml was altered to use the default PostgreSQL port instead of 15432. Also updated Dockerfile to grant execute permission to the new script and use it as the entrypoint.